+++ R/leaves.R
@@ -46,7 +46,7 @@ function(.tree)
 leaves.rtree <- # :: Tree -> [Node]
 function(.tree)
 {
-  if (is.empty(.tree)) {
+  if (is_empty(.tree)) {
     return(list())
   }
   if (length(.tree)==0L) {