+++ NEWS.md
@@ -1,1 +1,4 @@
-# oak 0.2.1
+# oak 0.2.3
+
+* 'is.empty()' is no longer reexported from package 'bazar', 
+and an internal function 'is_empty()'  is now used.